+++ src/Router.js
@@ -32,24 +32,21 @@ class Router extends Dispatcher {
     return this._pathCurrent
   }
 
-  createRoute (path, options = {}) {
-    // Create new route.
+  addRoute (path, options = {}) {
+    // Create routes test function.
     const regexp = pathToRegexp(path, [], options)
-    const match = (_path) => {
+    const test = (_path) => {
       return regexp.test(_path)
     }
 
     // Add route to list.
-    this._routes[path] = match
+    this._routes[path] = test
 
-    // Dispatch create event.
-    this.dispatch('created', {
+    // Dispatch add event.
+    this.dispatch('added', {
       route: path,
       router: this,
     })
-
-    // Return route.
-    return path
   }
 
   removeRoute (path) {
@@ -63,39 +60,47 @@ class Router extends Dispatcher {
   push (path) {
     // Remove base url, if present.
     const pathNew = path.replace(this._options.basePath, '')
-    const pathIsNew = this._pathCurrent !== pathNew
-    this._pathCurrent = pathNew
+    if (this._pathCurrent === pathNew) {
+      return true
+    }
 
     // Find matching routes.
+    let routeNew = null
     for (const routePath in this._routes) {
       const match = this._routes[routePath]
-      if (!match(this._pathCurrent)) {
+      if (!match(pathNew)) {
         continue
       }
 
-      this._routeCurrent = routePath
+      routeNew = routePath
       break
     }
 
-    if (pathIsNew) {
-      // Update page history if options set and window global exists.
-      if (this._options.updateHistory && typeof window !== 'undefined') {
-        // Construct url.
-        const url = path.includes(this._options.basePath) ? path : this._options.basePath + path
-        // Check if url is not current url.
-        if (url !== window.history.location) {
-          // Add path to history.
-          window.history.pushState(null, window.document.title, url)
-        }
-      }
-
-      // Dispatch event on router.
-      this.dispatch('pushed', {
-        path: this._pathCurrent,
-        route: this._routeCurrent,
-        router: this,
-      })
+    if (!routeNew) {
+      return false
     }
+    this._pathCurrent = pathNew
+    this._routeCurrent = routeNew
+
+    // Update page history if options set and window global exists.
+    if (this._options.updateHistory && typeof window !== 'undefined') {
+      // Construct url.
+      const url = path.includes(this._options.basePath) ? path : this._options.basePath + path
+      // Check if url is not current url.
+      if (url !== window.history.location) {
+        // Add path to history.
+        window.history.pushState(null, window.document.title, url)
+      }
+    }
+
+    // Dispatch event on router.
+    this.dispatch('pushed', {
+      path: this._pathCurrent,
+      router: this,
+      route: this._routeCurrent,
+    })
+
+    return true
   }
 }