+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## 1.0.0
 
+- Added documentation.
+- Added example.
 - Added Router export in root for standalone usage.
 - Added `getPath` and `removeRoute` functions to `Router` class.
-- Changes event names in router from `create` to `created` and `push` to `pushed`.
+- Changed attribute names of router from `onStart` to `onStarted`.
+- Changes event names in router from `create` to `added` and `push` to `pushed`.
 - Fixed updateHistory attribute of `router` component.
 - Removed `Route` class.
 - Reduced build size.